(TEST) add .long() to target in CE loss

+++ nitorch/nn/training/training.py
@@ -994,7 +994,7 @@ class SegGANTrainer:
             loss_adv_d = -torch.mean(real_valid) + torch.mean(fake_valid) + self.lambda_gp * grad_pen
 
             # domain
-            loss_dom_d = self.domain_loss(real_class, batch_s_met)
+            loss_dom_d = self.domain_loss(real_class, batch_s_met.long())
 
             # repeat for target -> source
             trans_s_img = self.model(image=batch_t_img, meta=batch_t_met,
